requirements.txt
cachetools==5.3.3
fastapi==0.110.0
fastui==0.5.2
pydantic==2.6.4
pydantic-settings==2.2.1
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
sqlalchemy==2.0.29
uvicorn==0.28.0
